Added comments and test cases
Added comments to the parser and 2 additional test scl files. Test1 and Test2 are the same, but I made some errors and commented on what I did. The comment lines will have to be removed, and the error will have to be corrected in order for the file to parse properly.